pyproject.toml
[tool.poetry]
name = "infobase"
version = "0.0.1"
description = ""
authors = ["desmond.zhan <desmond.zhan@cloudwise.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "infobase"}]

[[tool.poetry.source]]
name='tsinghua'
url="https://pypi.tuna.tsinghua.edu.cn/simple"
priority="default"

[tool.poetry.dependencies]
python = "^3.11.0"
langchain = "^0.0.317"
langserve = { version = ">=0.0.6", extras = ["server"] }
tiktoken = ">=0.4.0"
openai = "^0.27.8"
fastapi = "^0.96.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
gitpython = "^3.1.40"
markdown = "^3.5"
pymilvus = "^2.3.1"
transformers = "^4.34.1"
chromadb = "^0.4.14"
sentence-transformers = "^2.2.2"
tree-sitter = "^0.20.2"
python-dotenv = "^1.0.0"
whoosh = "2.7.4"
poetry = "^1.7.1"
py2neo = "^2021.2.4"
traceloop-sdk = "^0.21.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.4.1"
ruff = "^0.0.278"
black = "^23.7.0"
syrupy = "^4.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort
]

[tool.mypy]
ignore_missing_imports = "True"
disallow_untyped_defs = "True"

[tool.pytest.ini_options]
# --strict-markers will raise errors on unknown marks.
# https://docs.pytest.org/en/7.1.x/how-to/mark.html#raising-errors-on-unknown-marks
#
# https://docs.pytest.org/en/7.1.x/reference/reference.html
# --strict-config       any warnings encountered while parsing the `pytest`
#                       section of the configuration file raise errors.
#
# https://github.com/tophat/syrupy
# --snapshot-warn-unused    Prints a warning on unused snapshots rather than fail the test suite.
addopts = "--strict-markers --strict-config --durations=5 --snapshot-warn-unused"
